+++ .gitignore
@@ -12,6 +12,7 @@ dist
 dist-ssr
 *.local
 .env
+.env.local
 
 # Editor directories and files
 .vscode/*
