+++ public/index.html
@@ -5,21 +5,20 @@
   <title>Void Lounge</title>
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <meta name="theme-color" content="#0a0a1a" />
-  <link rel="stylesheet" href="/styles.css" />
-  
+  <link rel="stylesheet" href="styles.css" />
 </head>
 <body>
-  <!-- Background Particles -->
+  <!-- 🌀 Background Particles -->
   <canvas id="particles"></canvas>
 
-  <!-- Floating Dashboard Icon -->
+  <!-- 👁 Dashboard Toggle -->
   <div id="dashboard-toggle">👁</div>
   <div id="user-dashboard" class="hidden"></div>
 
-  <!-- Chat Messages -->
+  <!-- 💬 Message Container -->
   <div id="message-container"></div>
 
-  <!-- Chat Input Form -->
+  <!-- 📝 Chat Form -->
   <form id="chat-form" autocomplete="off">
     <input
       type="text"
@@ -29,12 +28,17 @@
     />
   </form>
 
-  <!-- Background Ambience Audio -->
+  <!-- 🎵 Ambient Audio -->
   <audio id="ambience" autoplay loop>
-    <source src="https://cdn.pixabay.com/download/audio/2022/02/23/audio_16728c7a56.mp3?filename=ambient-105276.mp3" type="audio/mpeg" />
+    <source
+      src="https://cdn.pixabay.com/download/audio/2022/02/23/audio_16728c7a56.mp3?filename=ambient-105276.mp3"
+      type="audio/mpeg"
+    />
   </audio>
 
-  <!-- App Logic -->
-  <script type="module" src="/main.js"></script>
+  <!-- ✅ Load main logic after DOM is ready -->
+  <script type="module" src="main.js"></script>
 </body>
 </html>
+
+
